moving specs to request specs, divide feature_edit_spec + map_layers_spec
