+++ spec/features/login_spec.rb
@@ -1,32 +1,16 @@
 require "rails_helper"
 
+# The login and logout flow itself is server side, see the request spec
+# spec/requests/sessions_controller_spec.rb. Only the profile dropdown needs a browser.
 describe "Login" do
+  let(:user) { create(:user) }
+
   before do
-    visit login_path
-  end
-
-  it "shows login options" do
-    expect(page).to have_button("Developer Login")
-  end
-
-  it "login via developer provider" do
-    click_button("Developer Login")
-    expect(page).to have_text("User Info")
-    fill_in "name", with: "Test User"
-    fill_in "email", with: "test@mapforge.org"
-    click_button("Sign In")
-    expect(page).to have_current_path(my_path)
-    expect(User.count).to eq 1
-  end
-
-  it "destroys session on logout" do
-    click_button("Developer Login")
-    fill_in "name", with: "Test User"
-    fill_in "email", with: "test@mapforge.org"
-    click_button("Sign In")
-    expect(page).to have_current_path(my_path)
-    expect(User.count).to eq 1
+    allow_any_instance_of(ApplicationController).to receive(:session).and_return({ user_id: user.id })
     visit maps_path
+  end
+
+  it "offers logout in the profile menu" do
     find(".profile-image").click
     click_link("Logout")
     expect(page).to have_current_path(root_path)
