+++ spec/features/maps_my_list_spec.rb
@@ -1,5 +1,7 @@
 require "rails_helper"
 
+# The list content and the filter are server rendered, see the request spec
+# spec/requests/maps_controller_spec.rb. Only the delete confirmation needs a browser.
 describe "Map List" do
   let(:user) { create(:user) }
 
@@ -7,10 +9,6 @@ describe "Map List" do
     allow_any_instance_of(ApplicationController).to receive(:session).and_return({ user_id: user.id })
     create(:map, owners: [ user ])
     visit my_path
-  end
-
-  it "shows private links to maps" do
-    expect(page).to have_selector(:xpath, "//a[@href='/m/#{user.owned_maps.first.private_id}']")
   end
 
   it "can delete own map" do
@@ -21,21 +19,5 @@ describe "Map List" do
     end
     expect(page).not_to have_css(".map-preview")
     expect(user.reload.owned_maps.count).to eq 0
-  end
-
-  context "filter list" do
-    before {
-      [ create(:map, owners: [ user ], name: "Map1"),
-        create(:map, owners: [ user ], name: "Map2") ]
-    }
-
-    it "searches in map names" do
-      fill_in "search", with: "Map1"
-      find_field("search").send_keys(:enter)
-
-      expect(page).to have_selector(".map-preview", count: 1)
-      expect(page).to have_text("Map1")
-      expect(page).not_to have_text("Map2")
-    end
   end
 end
