+++ spec/requests/application_controller_spec.rb
@@ -10,4 +10,18 @@ describe ApplicationController do
       expect(res).to have_http_status(200)
     end
   end
+
+  describe "#set_gettext_locale" do
+    it "translates the page into a requested locale" do
+      get root_path(locale: "de")
+      expect(response.body).to include('lang="de"')
+      expect(response.body).to include("Gestalte deine eigene Karte")
+    end
+
+    it "falls back to the default locale on an unknown locale" do
+      get root_path(locale: "xx")
+      expect(response.body).to include('lang="en"')
+      expect(response.body).to include("Create your own map")
+    end
+  end
 end
